BottomSheet now has an active bin Iconbutton + file structure
