+++ lib/src/features/menu/models/Section.dart
@@ -1,4 +1,4 @@
-import 'Drink.dart';
+import 'drink.dart';
 
 class Section {
   final String title;
